requirements.txt
asgiref==3.5.0
Django==4.0.4
djangorestframework==3.13.1
django-etc==1.3.1
django-hitcount==1.3.5
Pillow==9.1.0
sqlparse==0.4.2
tzdata==2022.1